Updated BookStore and BookStoreTest to use decimal for money

+++ exercises/book-store/BookStore.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 public static class BookStore
 {
-    public static double Total(IEnumerable<int> books)
+    public static decimal Total(IEnumerable<int> books)
     {
         throw new NotImplementedException("You need to implement this function.");
     }
